email.js tls fix
added condition to check if the port was 587 or 25 to disable the secure variable
as mentioned in the nodemailer doc:

secure – if true the connection will use TLS when connecting to server. If false (the default) then TLS is used if server supports the STARTTLS extension. In most cases set this value to true if you are connecting to port 465. For port 587 or 25 keep it false

https://nodemailer.com/smtp/#:~:text=secure%20%E2%80%93%20if%20true%20the%20connection%20will%20use,For%20port%20587%20or%2025%20keep%20it%20false